Moves components to components folder, removes logic from container #1074
